layout/style/test/test_flexbox_align_self_auto.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=696253
-->
<head>
  <meta charset="utf-8">
  <title>Test behavior of 'align-self:auto' (Bug 696253)</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=696253">Mozilla Bug 696253</a>
<div id="display">
  <iframe id="iframe"></iframe>
</div>
<pre id="test">
<script type="application/javascript">
"use strict";

/** Test behavior of 'align-self:auto' (Bug 696253) **/

/*
 * This mochitest runs in an iframe so that we can selectively turn on the
 * flexbox about:config pref before its document is instantiated.
 *
 * See the iframe's source ("file_flexbox_align_self_auto.html") for the actual
 * test code and for more documentation.
 */

SimpleTest.waitForExplicitFinish();

ok(SpecialPowers.getBoolPref("layout.css.flexbox.enabled"),
   "expecting flexbox pref to be enabled by default");

document.getElementById("iframe").src = "file_flexbox_align_self_auto.html";

function finish() {
  SpecialPowers.clearUserPref("layout.css.flexbox.enabled");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
